Minor improvements to description refactor
