+++ webapp/app/assessment/components/section/components/descriptions/descriptions.js
@@ -4,10 +4,9 @@ import { useSelector } from 'react-redux'
 
 import * as ObjectUtils from '@common/objectUtils'
 
-import NationalDataDescriptions from '@webapp/app/assessment/components/description/nationalDataDescriptions'
-import AnalysisDescriptions from '@webapp/app/assessment/components/description/analysisDescriptions'
-import CommentableDescription from '@webapp/app/assessment/components/description/commentableDescription'
-import useCountryIso from '@webapp/components/hooks/useCountryIso'
+import NationalDataDescriptions from '@webapp/app/components/description/nationalDataDescriptions'
+import AnalysisDescriptions from '@webapp/app/components/description/analysisDescriptions'
+import CommentableDescription from '@webapp/app/components/description/commentableDescription'
 import useI18n from '@webapp/components/hooks/useI18n'
 
 const Descriptions = props => {
@@ -21,17 +20,12 @@ const Descriptions = props => {
     ObjectUtils.isFunction(analysisAndProcessing) ? analysisAndProcessing(state) : analysisAndProcessing
   )
 
-  const countryIso = useCountryIso()
   const i18n = useI18n()
 
   return (
     <>
-      {useNationalData && (
-        <NationalDataDescriptions section={sectionName} countryIso={countryIso} disabled={disabled} />
-      )}
-      {useAnalysisAndProcessing && (
-        <AnalysisDescriptions section={sectionName} countryIso={countryIso} disabled={disabled} />
-      )}
+      {useNationalData && <NationalDataDescriptions section={sectionName} disabled={disabled} />}
+      {useAnalysisAndProcessing && <AnalysisDescriptions section={sectionName} disabled={disabled} />}
       {introductoryText && (
         <CommentableDescription
           section={sectionName}
